Exception handling - delete

+++ src/main/java/com/service/course/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.service.course.entities.User;
 import com.service.course.repositories.UserRepository;
+import com.service.course.services.exceptions.DatabaseException;
 import com.service.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -34,7 +36,18 @@ public class UserService {
 
 
     public void delete(Long id) {
-        repository.deleteById(id);
+
+        boolean exists = repository.existsById(id);
+
+        if(!exists) {
+            throw new ResourceNotFoundException(id);
+        }
+
+        try {
+            repository.deleteById(id);
+        } catch(DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
 
@@ -45,11 +58,11 @@ public class UserService {
         return repository.save(entity);
     }
 
+
     private void updateData(User entity, User obj) {
         entity.setName(obj.getName());
         entity.setEmail(obj.getEmail());
         entity.setPhone(obj.getPhone());
     }
 
-    
 }
